Add .uniq for rendering surveys, studies and data streams. Edit the survey panel for researcher.

+++ app/models/version.rb
@@ -1,11 +1,11 @@
-class Version 
-  include Mongoid::Document 
+class Version
+  include Mongoid::Document
 
   store_in collection: 'dataPoint', database: 'omh'
-  
-  field :major, type: Float
-  field :minor, type: Float 
-  
+
+  field :major, type: Integer
+  field :minor, type: Integer
+
   embedded_in :schema_id, :inverse_of => :version
 
-end 
+end
